Remove dependency on package `curl`
sagit:puppet already ensures the presence of package 'curl',
leading to puppet-rocket failing with the error:

  `Package[curl] is already declared …; cannot redeclare`

The initial declaration is in

  sagit:puppet/manifests/default_node.pp:9
  `ensure_packages($params[packages])`

with the package params being defined in

  hieradata/common.yaml:53
  `- curl`

AFAICT, puppet-rocket doesn't actually use curl; it uses wget.
So we don't need the package at all!

Solution: just remove the dependency entirely.