[AR-2653] Ensure PerimeterX doesn't block calls to SAPI
When a user pastes in a URL, RocketChat fetches the URL
to render a Link Preview.  There are two problems here:

1. PerimeterX sometimes CAPTCHAs the request
2. On some URLs, SAPI declines the request for unsubscribed users.

We have defined a custom HTTP header by which RocketChat can
authenticate itself to both PerimeterX and SAPI:
   - I've patched RocketChat to send that header to seekingalpha.com
     (and its subdomains) exclusively
   - PerimeterX whitelists requests with that header key/value
   - SAPI can check for that header when needed

n.b. "Speakeasy" refers to a secret clubhouse, entrance to which
is obtained by saying a passphrase.